Move tsd assert api in is own directory
