+++ source/index.ts
@@ -1,4 +1,4 @@
 import tsd from './lib';
 
-export * from './lib/assertions/assert';
+export * from './lib/assertions/tsd/assert';
 export default tsd;
