+++ source/lib/assertions/tsd/handlers/identicality.ts
@@ -1,6 +1,6 @@
 import {CallExpression, TypeChecker, TypeFlags} from '@tsd/typescript';
-import {Diagnostic} from '../../interfaces';
-import {makeDiagnostic} from '../../utils';
+import {Diagnostic} from '../../../interfaces';
+import {makeDiagnostic} from '../../../utils';
 
 /**
  * Asserts that the argument of the assertion is identical to the generic type of the assertion.
