+++ source/lib/assertions/tsd/handlers/strict-assertion.ts
@@ -1,6 +1,6 @@
 import {CallExpression, TypeChecker} from '@tsd/typescript';
-import {Diagnostic} from '../../interfaces';
-import {makeDiagnostic} from '../../utils';
+import {Diagnostic} from '../../../interfaces';
+import {makeDiagnostic} from '../../../utils';
 
 /**
  * Performs strict type assertion between the argument if the assertion, and the generic type of the assertion.
